+++ pyscript/core/utils/path.py
@@ -1,28 +1,27 @@
-from os import getcwd as osgetcwd
-from os.path import sep, abspath as osabspath, normpath as osnormpath, splitext, basename
-
 from .string import normstr
+
+import os
 
 def getcwd() -> str:
     try:
-        return osgetcwd()
+        return os.getcwd()
     except:
         return '.'
 
 def abspath(path: str) -> str:
     try:
-        return osabspath(path)
+        return os.path.abspath(path)
     except:
         return path
 
 def base(path: str) -> str:
-    return splitext(basename(path))[0]
+    return os.path.splitext(os.path.basename(path))[0]
 
 def extension(path: str) -> str:
-    return splitext(basename(path))[1]
+    return os.path.splitext(os.path.basename(path))[1]
 
 def normpath(*paths, absolute: bool = True) -> str:
-    path = osnormpath(sep.join(map(normstr, paths)))
+    path = os.path.normpath(os.path.sep.join(map(normstr, paths)))
     return abspath(path) if absolute else path
 
 def get_name_from_path(path) -> str: