+++ changelog.md
@@ -1,6 +1,11 @@
 # Change Log
 
-## [1.12.6] - 14/03/2026
+## [1.12.7] - 18/03/2026
+
+### Added
+- Updated history features in the advanced shell (_REPL_) with `pygments`.
+- New node for the constant keyword `__debug__`.
+- Wrap around pointer and garbage collection mechanism in the `brainfuck` module.
 
 ### Fixed
 - **Fixed some bugs**.